pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=12.2.0",
    "pydantic-settings>=2.14.0",
    "requests>=2.33.1",
    "tkintermapview>=1.29",
]
